Datasets used in Spatial Clustering Python file
The 201806sum_all_data file contains data gathered from CitiBike, which tracked when bikes are rented and returned to their bike stations. The data was analyzed to represent over 760 entries with four columns: the station ID, GPS location of the bike stations, and their rent and return traffic for the given time period. The adjacency file is a matrix that indicates whether a particular bike station is a neighbor of another bike station.